Enabling autoupdate
If there is an CS:Go update available, the server will:
- wait untill the current game is over
OR
- check if the server is idle
before installing the update and restarting the game.
The script used to update the game is created within the Dockerfile.

Important: The srcds_run is broken at the moment, so we are repairing it by an simple RUN sed -i 's/steam.sh/steamcmd.sh/g' /home/cs/csgo_server/srcds_run